AVANCES

+++ movilidad-sostenible/usuario-service/src/main/java/com/movilidadsostenible/usuario/services/UserService.java
@@ -11,13 +11,13 @@ public interface UserService {
     void delete(String uid);
     Optional<User> byId(String uidUser);
 
-    Integer getBalance(String uidUser);
-    Integer addBalance(String uidUser, Integer amount);
-    Integer subtractBalance(String uidUser, Integer amount);
+    Double getBalance(String uidUser);
+    Double addBalance(String uidUser, Double amount);
+    Double subtractBalance(String uidUser, Double amount);
 
     // Devuelve true si el usuario NO puede viajar (bloqueado), false si SÍ puede
     boolean isUserBlockedForTravel(String uidUser);
 
     // Cobra el viaje aplicando reglas de suscripción.
-    void chargeTravel(Integer totalTripValue, Integer excessMinutes, String uidUser);
+    void chargeTravel(Double totalTripValue, Integer excessMinutes, String uidUser);
 }
